frontend/requirements.txt
streamlit==1.56.0
requests==2.33.1

langsmith>=0.1.98,<0.2.0
python-dotenv>=1.0.1